refactor: share one read-only LaTeX view and exclude the tools table from duplication checks
The TikZ preview copied the import pane's read-only CodeMirror mount, so
both now render through one ReadOnlyLatex component (the import viewer
keeps its gutter and test id). The assistant tools table is a list of
same-shaped rows that copy-paste detection reads as a clone of itself, so
it joins the other data tables in sonar.cpd.exclusions; its other rules
still run.
