+++ sonar-project.properties
@@ -26,7 +26,7 @@ sonar.coverage.exclusions=scripts/**,src/developer/**,src/lib/e2e-import-registr
 # tables is guarded by catalog-parity.test.ts and research-seed-catalog.test.ts
 # instead, which check the entries against the Rust catalog and against the
 # fixture directories on disk.
-sonar.cpd.exclusions=packages/ai-core/src/providers.ts,crates/oleafly-agent/src/provider.rs,src/developer/research-seed-catalog.ts
+sonar.cpd.exclusions=packages/ai-core/src/providers.ts,crates/oleafly-agent/src/provider.rs,src/developer/research-seed-catalog.ts,src/components/ai/AiToolsList.tsx
 
 # The existing Rust CI job already runs Clippy with warnings denied. Running it
 # again inside the scanner would duplicate work without adding a distinct gate.
